feat: api pokémon documentée et dto séparés
